Moved create glossary request to a new glossary namespace, updated references to create request, reverted change in Gemfile dependencies
